requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
transformers==4.37.2
torch>=2.6.8
numpy<2.0.0
openai==1.51.0
httpx==0.27.0
beautifulsoup4==4.12.3
newspaper3k==0.2.8
sentence-transformers==2.2.2

